Assignment 1/Part B/resume_analyzer/requirements.txt
Flask==2.0.1
spacy==3.1.3
python-docx==0.8.11
pdfminer.six==20201018